server-side cache implemented

+++ src/services/postgres/NotesService.js
@@ -6,9 +6,10 @@ const { nanoid } = require('nanoid')
 const { mapDBToModel } = require('../../utils')
 
 class NotesService {
-  constructor (collaborationsService) {
+  constructor (collaborationsService, cacheService) {
     this._pool = new Pool()
     this._collaborationsService = collaborationsService
+    this._cacheService = cacheService
   }
 
   async verifyNoteOwner (id, owner) {
@@ -63,19 +64,28 @@ class NotesService {
       throw new InvariantError('Catatan gagal ditambahkan')
     }
 
+    await this._cacheService.delete(`notes:${owner}`)
     return successId
   }
 
   async getNotes (owner) {
-    const query = {
-      text: `SELECT notes.* FROM notes
+    try {
+      const result = await this._cacheService.get(`notes:${owner}`)
+      return JSON.parse(result)
+    } catch (error) {
+      const query = {
+        text: `SELECT notes.* FROM notes
       LEFT JOIN collaborations ON collaborations.note_id = notes.id
       WHERE notes.owner = $1 OR collaborations.user_id = $1
       GROUP BY notes.id`,
-      values: [owner]
+        values: [owner]
+      }
+      const result = await this._pool.query(query)
+      const mappedResult = result.rows.map(mapDBToModel)
+
+      await this._cacheService.set(`notes:${owner}`, JSON.stringify(mappedResult))
+      return mappedResult
     }
-    const result = await this._pool.query(query)
-    return result.rows.map(mapDBToModel)
   }
 
   async getNoteById (id) {
@@ -97,7 +107,7 @@ class NotesService {
   async editNoteById (id, { title, body, tags }) {
     const updatedAt = new Date().toISOString()
     const query = {
-      text: 'UPDATE notes SET title=$1, body=$2, tags=$3, updated_at=$4 WHERE id = $5 RETURNING id',
+      text: 'UPDATE notes SET title=$1, body=$2, tags=$3, updated_at=$4 WHERE id = $5 RETURNING id, owner',
       values: [title, body, tags, updatedAt, id]
     }
     const result = await this._pool.query(query)
@@ -105,11 +115,14 @@ class NotesService {
     if (!result.rows.length) {
       throw new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan')
     }
+
+    const { owner } = result.rows[0]
+    await this._cacheService.delete(`notes:${owner}`)
   }
 
   async deleteNoteById (id) {
     const query = {
-      text: 'DELETE FROM notes WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM notes WHERE id = $1 RETURNING id, owner',
       values: [id]
     }
 
@@ -118,6 +131,9 @@ class NotesService {
     if (!result.rows.length) {
       throw new NotFoundError('Catatan gagal dihapus. Id tidak ditemukan')
     }
+
+    const { owner } = result.rows[0]
+    await this._cacheService.delete(`notes:${owner}`)
   }
 }
